Refactor: Bring Export Product modal to a different component
